+++ sphinx/source/changelog.rst
@@ -4,6 +4,8 @@ Changelog
 1.0.2b1
 -------
 * Added a second source for the Skool Daze TZX file to `images.ini`
+* Added the ``--package-dir`` command line option (for showing the path to the
+  pyskool package directory)
 * Removed the documentation sources from the Pyskool distribution (they can be
   obtained from GitHub_)